requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
branca==0.8.1
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
deep-sort-realtime==1.3.2
dnspython==2.7.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
folium==0.20.0
fonttools==4.59.1
fsspec==2025.7.0
future==1.0.0
geocoder==1.38.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
narwhals==2.1.2
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydeck==0.9.1
pymongo==4.14.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.1
streamlit-folium==0.25.1
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.184
ultralytics-thop==2.0.16
urllib3==2.5.0
Werkzeug==3.1.3
xyzservices==2025.4.0
